+++ client/package.json
@@ -6,7 +6,6 @@
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.11.2",
     "axios": "^0.21.0",
-    "classnames": "^2.3.1",
     "moment": "^2.29.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
