--- client/src/themes/hooks/index.js
@@ -1,7 +1,0 @@
-import useButtonStyles from "./buttons";
-import useFormStyles from "./Forms";
-
-export {
-    useButtonStyles,
-    useFormStyles
-};
